Crate actix_http_test

Various helpers for Actix applications to use during testing.

Structs

TestServer

Test server controller

Functions

test_server

Start test server.

test_server_with_addr

Start test server on an existing address binding.

unused_addr

Get a localhost socket address with random, unused port.